Restructure to confrom to the component diagram in the TO
