+++ application_logic/dashboard_controller/devices_controller.py
