+++ application_logic/sensor_controller/nfctest.py
