+++ config.py
@@ -1,0 +1,5 @@
+class Config(object):
+    DEBUG = True
+
+class Config_Prod(Config):
+    DEBUG = False
