--- dashboard/config.py
@@ -1,11 +1,0 @@
-class Config(object):
-    DEBUG = True
-
-    # email server
-    MAIL_SERVER = 'smtp.gmail.com'
-    MAIL_PORT = 465
-    MAIL_USE_SSL = True
-    MAIL_USERNAME = ''
-    MAIL_PASSWORD = ''
-
-    ADMINS = ['maarten.deklerk@gmail.com']
