--- database.sql
@@ -1,35 +1,0 @@
-BEGIN TRANSACTION;
-CREATE TABLE IF NOT EXISTS 'water_damages' (
-	'id'	INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-	'date'	DATETIME NOT NULL,
-	'devices_id'	INTEGER NOT NULL UNIQUE,
-	FOREIGN KEY('devices_id') REFERENCES 'devices'('id')
-);
-CREATE TABLE IF NOT EXISTS 'users' (
-	'id'	INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-	'email'	TEXT NOT NULL UNIQUE,
-	'password'	TEXT NOT NULL
-);
-CREATE TABLE IF NOT EXISTS 'rooms' (
-	'id'	INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-	'name'	VARCHAR ( 50 ) NOT NULL,
-	'homes_id'	INTEGER NOT NULL,
-	FOREIGN KEY('homes_id') REFERENCES 'homes'('id')
-);
-CREATE TABLE IF NOT EXISTS 'ice' (
-	'id'	INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-	'name'	VARCHAR ( 255 ) NOT NULL,
-	'type'	CHAR ( 1 ) NOT NULL
-);
-CREATE TABLE IF NOT EXISTS 'homes' (
-	'id'	INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-	'location'	VARCHAR ( 100 ) NOT NULL
-);
-CREATE TABLE IF NOT EXISTS 'devices' (
-	'id'	INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-	'name'	VARCHAR ( 50 ) NOT NULL,
-	'status'	CHAR ( 1 ) NOT NULL,
-	'rooms_id'	INTEGER NOT NULL,
-	FOREIGN KEY('rooms_id') REFERENCES 'rooms'('id')
-);
-COMMIT;
