+++ static/img/sensor-green.png
